Remove mainview and put all the view configuration in controller
